requirements.txt
asgiref==3.8.1
Django==3.2.14
django-debug-toolbar==4.4.6
djangorestframework==3.12.2
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
playwright~=1.49.1
pytest~=8.3.4
bs4~=0.0.2
beautifulsoup4~=4.12.3
soupsieve~=2.6
pip~=24.3.1
wheel~=0.45.1
pyee~=12.0.0
setuptools~=75.6.0
pluggy~=1.5.0
exceptiongroup~=1.2.2
iniconfig~=2.0.0
certifi~=2024.12.14
python-slugify~=8.0.4
urllib3~=2.3.0
greenlet~=3.1.1
idna~=3.10
charset-normalizer~=3.4.1
packaging~=24.2
tomli~=2.2.1
requests~=2.32.3
pytest-playwright~=0.6.2